Replace powerpc64le with ppc64le where relevant
powerpc64le is not the default name, ppc64le is.

Historically powerpc64le was choosen, but it is now causing
repeated issues every time freedesktop-sdk needs to be used on that
platform because it's not the default name and it needs to be
overriden everywhere. This is a breaking change for users of
freedesktop-sdk on that platform because the 'target_arch' and
'bootstrap_build_arch' options now accept ppc64le and not
powerpc64le anymore.

powerpc64le is used by default by LLVM and GCC, but for
freedesktop-sdk it makes more sense to use ppc64le because that's
flatpak's and buildstream's default and Linux's `uname -m` on all
GNU/Linux distributions.