Add new reproducibility test in python
This is a rewrite of the reproducibility script in shell
to teach me what are the parts of buildstream that are
lacking for proper script integration.

This script uses the environment variable BST when defined
for the sole purpose of parsing the config file and
using the correct location for the artifact directory,
if not defined, we check in XDG_CONFIG_PATH/buildstream1.conf
and then in XDG_CONFIG_PATH/buildstream.conf in that order.

We tried to minimize the number of calls to buildstream
to speed up the process.

pipe the stderr to a different file to have only the script
output on the screen (userfull for debugging the script), if
you chose to run the script with stdout and stderr you will also
have to deal with buildstream messages.

if you need help, run the script with --help.