.gitlab-ci.yml: Merge publish stages
For 20.08.0 we were not able to publish the Docker image because an
error in gitlab CI marked the finish publish for flatpak as failed
even though it worked. And we could not restart it. The publishing
of snap and docker images depended on this stage to pass.

So instead we publish everything in the same time so we can retry
manually all failed steps.