tests/reproducible-test.bst: Test reproducibility on all elements
We do not follow build dependencies, so both components and output
have to be listed. Also base elements that are split through filter
have to be listed for the moment.

`sdk.bst` and `platform.bst` are explictly added as safety in case it
was forbidden in `components.bst`.  `oci/debug-oci.bst` is omitted due
to the time and size it takes to build it. Snaps are not added yet.