Copy dynamic loader debug file in SDK
This file is required by valgrind. Normally the file is in devel split
domain rather than debug. But due the spaghettis in our split rules for
the Flatpak images, this file was lost and not added to the SDK.

It is relatively difficult to keep the file in devel and not in debug
because of limitation in how BuildStream split rules work: we lose the
rules through compose elements, we cannot properly compose directly
within flatpak_image, we cannot filter have exceptions from globs in a
split domain.