Bootstrap Go
This will helpful to bootstrap new architectures which do not have a
binary release yet. For example RISC-V. Once it will hit the
PreBootstrap, we can remove our dependency to the binary.

It is not possible yet to properly cross-build go since when
target arch is the same as build, it does not properly use
build glibc for the build go. Full go with CGO enabled needs to be
built later out of bootstrap phase. See bug
https://github.com/golang/go/issues/25177