include/flags: compile with -fno-omit-frame-pointer for 64bit systems
When using "-01" in GCC and higher it also toggles the `-fomit-frame-pointer`
flag. This used to have a performance impact on i686 machines, and has
been historically why it was left on.

With x86_64 though the registers are part of the calling convention,
and there's virtually no benefit of disabling it. Another small side
effect could be, slightly bigger binary sizes.

On the other side it makes stack traces much harder to unwind and
profiling tools like perf and Sysprof can't function. This makes
it fairly hard to debug crashes as application developers would
need to rebuild the runtime (and possibly the downsteam runtime,
like gnome or kde ones) or rebuild all of their applications dependencies.

Close #1210