Add support to Vulkan implicit layers for GL extensions
Some components rely on Vulkan implicit layers. In particular, they
can be important on multi-GPU systems: many Vulkan games and
applications default to using whichever GPU is enumerated first, and
the way to achieve this in Vulkan is to re-order the devices in an
implicit layer (Mesa's device selection layer and/or NVIDIA's
Optimus layer).

Similar to what's already done for Vulkan explicit layers, this
commit adds support for exporting implicit layers from a GL
extension, so that the Mesa and NVIDIA driver extensions can export
their implicit layers.

Helps:
https://github.com/flathub/org.freedesktop.Platform.GL.nvidia/issues/35

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>