Include the mesa VK_LAYER_MESA_device_select implicit layer
If there are multiple GPUs available, how they are ordered is not always
how the user would have expected.

For example if there is a single dedicated GPU and the software based
lavapipe, it might happen that lavapipe gets chosen as the first GPU.
And it's quite common that applications and games just use the first GPU
as the default one.

By including the implicit Vulkan layer VK_LAYER_MESA_device_select, the
GPU ordering is expected to be more consistent and less unexpected
because it prioritizes the device that is the connection primary (if
there is an X or wayland connection) or the device which was the
boot_vga.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>