Use updated cache server
The new cache server uses Buildbarn[0] as a backend, rather than a
bst-artifact-server. It consists of a bb-storage[1] instance operating a
CAS server (for storing the underlying blobs) and a bb-remote-asset[2]
instance to provide a mapping from BuildStream artifact cache keys to
the artifacts themselves (albeit technically with some indirection
handled by BuildStream). These are placed behind a Traefik[3] instance
which provides routing and TLS termination - it has a built in ACME
client which will keep a certificate from Let's Encrypt up to date.

The curious can see the deployment configuration in the `artifacts`
repository in the infra group[4].

[0]: https://github.com/buildbarn
[1]: https://github.com/buildbarn/bb-storage
[2]: https://github.com/buildbarn/bb-remote-asset
[3]: https://docs.traefik.io
[4]: https://gitlab.com/freedesktop-sdk/infrastructure/artifacts/-/merge_requests/1