components/u-boot-qemu.bst: Fix partition signature in EFI
Systemd Boot sets an EFI variable with the GUID of the EFI system
parition it was loaded from. Then later, Systemd will read that
variable to find any partition with the root GUID that is on the same
drive as the EFI system partition. Then that root partition gets a
symlink as /dev/gpt-auto-root. If no root= parameter was given as
kernel parameter Systemd from an initramfs would mount
/dev/gpt-auto-root.

Unfortunately when querying the GUID of the system partition in
systemd-boot, U-Boot gives instead the GUID of the partition *table*.

So systemd-boot would not find /dev/gpt-auto-root.

I believe this is a U-Boot bug and the patch should get upstreamed.