Remove non-deterministic data from gnupg
Some "*.po" files were automatically generated, and
included the current date in "PO-Revision-Date". That
date ended up in the compiled "*.mo" files making `gnupg.bst`
build not deterministic. This is due to `xgettext`
and `msginit` not producing reproducible outputs[0]

There were some fixes already in `gettext` but they only
covered "PO-Revision-Date".

Additionally, `gnupg` includes a script to filter out similar
things, but that was only handling "POT-Creation-Date".

Now we build manually the affected "*.po" files so that we
are able to remove that information in the middle of the
build process.

[0]: https://savannah.gnu.org/bugs/?59658