bootstrap/coreutils.bst: Force cross compilation configuration into cache
Because glibc 2.33 binaries cannot run on glibc 2.32, configure thinks
that lots of thing misbehave on target OS. But it is just that the
conf test binaries cannot run. So we have to for the configuration.

When --host= is the same as --build=, then autoconf thinks that it is
not cross-compilation, so it does it still tries to run binaries for
target OS which has an incompatible glibc.

This was causing `cp -a` to fail on symlinks because gnulib was trying
to work around issues that did not exist.