utils/test_reproducibility.py: Fix overall return status.
This fixes the overall exit status of the script to be 0 in the case
that everything was found to be reproducible.

It appears that there may have been a change in what was returned by
bst_check_reproducibility_v2(), where it used to return a list but was
changed to return a dictionary, and it appears that handle_results() was
not updated to take that into consideration when evaluating the overall
success.