components/linux.bst: Use olddefconfig instead of oldconfig
After changing configuration with `scripts/config` after `defconfig`, we
need resolve selects, set default values for unlocked configurations,
etc. We need to do this otherwise `make` will try to ask us questions
on what to do about configurations that were not answered.

Unfortunately `yes '' | make oldconfig` does not work correctly in
case of optional "choice". In that case the prompt will first ask if
we want to answer the choice, and `yes ''` will be the same as
"no". So the default choice is not taken, or even worse, if we
selected the choice with `scripts/config`, it will be just remove that
configuration.

Using `make olddefconfig` instead does not suffer of this issue.