gstreamer-plugins-base.bst: Add patch to fix a gl/wayland assertion
As freedesktop-sdk is shipping gstreamer 1.18 currently, it didn't
get a patch that fixed the GL/Wayland glimagesink when using newer
libwayland-client. The GTK 4 widget factory currently crashes for
me because of that. So until gstreamer is bumped to a version with
this patch included, let's ship it as a patch to gst-plugins-base.

See https://gitlab.gnome.org/GNOME/gtk/-/issues/4137 for a GTK issue.
Patch is at https://gitlab.freedesktop.org/gstreamer/gst-plugins-base/-/merge_requests/1200