extensions/mesa/libdrm.bst: Disable core libdrm tests
Since the libdrm core is not actually built or packaged in the Mesa
extension, it shouldn't build its tests either. These tests are
already built as part of the libdrm core in the runtime
(components/libdrm.bst).

It's important to make this distinction because otherwise we would
see build failures when new API is added to the libdrm core along
with new tests. Since our core libdrm API is frozen, these tests
would fail to find the newly added symbols during link time, resulting
in build failures of the Mesa extension.

Unfortunately, core libdrm tests are not clearly marked as such, so
each test had to be reviewed to see if it relied upon driver-specific
API or only core API.

This commit modifies the existing patch to add the "with_libdrm"
variable to meson.build, which is then used to skip building the
relevant tests. It also removes any modifications to these tests'
build commands as they should never be built against the external
core libdrm library.