at-spi2-core: Enable dbus-daemon
Without either `dbus-daemon` or `dbus-broker`, at-spi2-core gets build
into something that never properly starts. Flatpak depends on
`org.a11y.Bus` and hangs for some seconds if it does not start
properly. So for Flatpak to work properly in VM images we need to add
`dbus.bst` in build dependencies so it detects `dbus-daemon`.