Update GDB to 11.2
The existing version has an unfortunate bug that makes the debugger more or less
useless when it is built with asserts. The fix was landed upstream here:
https://github.com/bminor/binutils-gdb/commit/2ccee230f83055af67cb34781179b24a4b9ef2f2
with the earliest release containing the fix being 11.1.

I went ahead and updated to 11.2 but I am not particularly well-initiated to the
intricacies of binutils, so I'm not sure if there's other components that need
to be updated at the same time; I was able to rebuild the SDK locally with this
patch, however