components/systemd-base.bst: Cleanup the sbat vendor entry
The id is supposed to be <upstream>.<vendor>. Having dots makes it not
obvious what part is upstream and what is vendor. Summary can use
capitals and spaces. URL is not a bug report URL, but more the main
page for the project. There was a '+' as suffix in the version tag
because of patches. We should not have that.