Add variables to libtheora.bst to build correctly using autogen.sh
In elements/components/libtheora.bst the autogen variable has been
redefined so that the variables --build and %{conf-args} are
passed to autogen.sh. This is because autogen.sh script doesn't
respect NOCONFIGURE. The make, and make-install variables have
been redefined. This is necessary becuase when  in %{build-dir}
the makefile generated by autogen.sh cannot be found without
passing a path to the makefile. Statements have been added to
generate and change into %{build-dir} to stay consistent with
other components. The source url has been changed to the
canonical source, and the extra trailing newline has been removed.