components/bash-config.bst: Load /etc/profile.d/*.sh from bashrc
Prior to this commit, the bash configuration we provide does not load
properly `/etc/profile.d`. There are several things not working
properly, for instance GNOME Console integration.

We also move most of the the skeleton into a common /etc/bashrc file
which can be updated by the systems. Whereas skeleton files
are copied only once.